PDF-RAG-chatbot/requirements.txt
streamlit
python-dotenv
requests
langchain
langchain-community
langchain-core
langchain-huggingface
sentence-transformers
qdrant-client
pymupdf4llm
unstructured
Pillow
nltk
